refactor: move recipe models to gibsonify api
Move models from recipe feature to  gibsonify_api, while resolving name conflicts.

BREAKING CHANGE:
