+++ lib/recipe/bloc/recipe_bloc.dart
@@ -3,14 +3,18 @@ import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
 import 'package:uuid/uuid.dart';
 
-import 'package:gibsonify/recipe/models/models.dart';
-import 'package:gibsonify/recipe/recipe.dart';
+import 'package:gibsonify_api/gibsonify_api.dart';
+import 'package:gibsonify_repository/gibsonify_repository.dart';
 
 part 'recipe_event.dart';
 part 'recipe_state.dart';
 
 class RecipeBloc extends Bloc<RecipeEvent, RecipeState> {
-  RecipeBloc() : super(const RecipeState()) {
+  final GibsonifyRepository _gibsonifyRepository;
+
+  RecipeBloc({required GibsonifyRepository gibsonifyRepository})
+      : _gibsonifyRepository = gibsonifyRepository,
+        super(const RecipeState()) {
     on<RecipeAdded>(_onRecipeAdded);
     on<RecipeDeleted>(_onRecipeDeleted);
     on<RecipeNameChanged>(_recipeNameChanged);
@@ -31,6 +35,7 @@ class RecipeBloc extends Bloc<RecipeEvent, RecipeState> {
     on<IngredientMeasurementUnitChanged>(_onIngredientMeasurementUnitChanged);
     on<IngredientSizeChanged>(_onIngredientSizeChanged);
     on<IngredientSizeNumberChanged>(_onIngredientSizeNumberChanged);
+    // on<RecipesSaved>(_onRecipesSaved); // TODO:
   }
 
   void _onRecipeAdded(RecipeAdded event, Emitter<RecipeState> emit) {
@@ -437,4 +442,9 @@ class RecipeBloc extends Bloc<RecipeEvent, RecipeState> {
 
     emit(state.copyWith(recipes: recipes));
   }
+
+  // TODO:
+  // void _onRecipesSaved(RecipesSaved event, Emitter<RecipeState> emit) {
+  //   List<Recipe> recipes = List.from(state.recipes);
+  // }
 }
