--- lib/recipe/models/models.dart
@@ -1,2 +1,0 @@
-export 'recipe_ingredient.dart';
-export 'recipe_item.dart';
