+++ lib/recipe/recipe.dart
@@ -1,4 +1,3 @@
 export 'bloc/recipe_bloc.dart';
-export 'models/models.dart';
 export 'view/view.dart';
 export 'widgets/widgets.dart';
