+++ packages/gibsonify_api/lib/src/models/food_item.dart
@@ -1,7 +1,10 @@
 import 'package:equatable/equatable.dart';
 import 'package:formz/formz.dart';
-import 'package:gibsonify_api/gibsonify_api.dart';
 import 'package:uuid/uuid.dart';
+
+// TODO: This is a temporary fix to avoid Name conflicts when importing the api
+import 'recipe_ingredient.dart'
+    show Name, Description, MeasurementMethod, MeasurementUnit;
 
 class FoodItem extends Equatable {
   FoodItem(
@@ -101,7 +104,13 @@ class FoodItem extends Equatable {
         confirmed
       ];
 }
+// TODO: This has been temporarily commented out to resolve name conflicts
+// The most sensible fix to this is probably getting rid of all Formz and
+// replacing them by strings only, as in 90% of the cases the only validation
+// is checking if it is not empty, and the rest can be added as custom
+// validation methods
 
+/*
 enum NameValidationError { invalid }
 
 // TODO: Investigate changing classes to be private (with leading underscore)
@@ -127,6 +136,7 @@ class Name extends FormzInput<String, NameValidationError> {
     return value?.isNotEmpty == true ? null : NameValidationError.invalid;
   }
 }
+*/
 
 enum TimePeriodValidationError { invalid }
 
@@ -193,6 +203,7 @@ class Source extends FormzInput<String, SourceValidationError> {
   }
 }
 
+/*
 enum DescriptionValidationError { invalid }
 
 class Description extends FormzInput<String, DescriptionValidationError> {
@@ -215,6 +226,7 @@ class Description extends FormzInput<String, DescriptionValidationError> {
         : DescriptionValidationError.invalid;
   }
 }
+*/
 
 enum PreparationMethodValidationError { invalid }
 
@@ -262,6 +274,8 @@ class PreparationMethod
   }
 }
 
+// TODO: allowed measurement methods are changed for not empty check only
+/*
 enum MeasurementMethodValidationError { invalid }
 
 class MeasurementMethod
@@ -297,7 +311,7 @@ class MeasurementMethod
         : MeasurementMethodValidationError.invalid;
   }
 }
-
+*/
 enum MeasurementValueValidationError { invalid }
 
 // TODO: Investigate changing this to a string or float
@@ -326,6 +340,8 @@ class MeasurementValue
   }
 }
 
+// TODO: allowed measurement units are changed for not empty check only
+/*
 enum MeasurementUnitValidationError { invalid }
 
 class MeasurementUnit
@@ -362,3 +378,4 @@ class MeasurementUnit
         : MeasurementUnitValidationError.invalid;
   }
 }
+*/
