+++ packages/gibsonify_api/lib/src/models/models.dart
@@ -1,2 +1,4 @@
 export 'food_item.dart';
 export 'gibsons_form.dart';
+export 'recipe_ingredient.dart';
+export 'recipe_item.dart';
